char.nl, nit_builtin.nl: Moved the definition of the `character' type from char.nl to nit_builtin.nl, since it is builtin in Mercury.
make_hlds.nl: Allow a type to be declared multiple times so long as it
is only defined once.  (This is so that you can declare a type in the
interface section of a module, and then define the type in the
implementation section.)

make_hlds.nl: Ensure that clauses are inserted into the hlds in the
correct order rather than in reverse order, so that the type-checker
checks them in the right order.