The various memory areas are now allocated dynamically, and one can specify their sizes on the comand line.
The layout of the memory areas is optimized for reducing conflicts
in primary caches. Any mrXY that are not real registers are #defined
to unreal_reg_XY. Unfortunately, one cannot control the layout of
these variables wrt the cache. #defining excess mrXY as unreal_regs[XY]
avoids conflicts but nevertheless slows down access considerably
due to the longer instruction sequences required.

On SVR4 machines, we now set up a redzone at the end of each area.
The sizes of these areas are configurable by options.

Labels used to be stored in a fixed-size table with linear search.
They are now stored in two hash tables, which can be lookup up
either by address or by name.

The timimg package now counts time in milliseconds and not microseconds,
thus allowing longer runs without overflow.