Merge in changes from version 4.5 of the Boehm conservative garbage collector.
Fix the Makefile so that if the -Wo,notail option (a work-around for a
bug in the DEC C compiler's tail optimization, apparently) does not
work, the error is ignored; this makes it work with gcc.

Change the Makefile so that there is a single rule for creating
libgc.so, rather than one for libgc.so for SunOS and Irix, and
another different one for the Alpha.