A new library file based on map.m, where the data-type is a list of things.
Estimated hours taken: _3___

A new library file based on map.m, where the data-type is a list of things.
This gives a vague one-to-many storage class.  It's based on map for
flexability, and doesn't achieve much more than map and list together, but
it is nice.  It is also completely untested.  I'm adding it now, because I
disappear for 12 months in 4 days time.

library/multi_map.m
	added this file.