Fix some bugs in which the failure continuation was not being set to
Estimated hours taken: 3

Fix some bugs in which the failure continuation was not being set to
`unknown' after a model_nondet goal.  We did it for model_nondet
disjunctions and calls, but not for the other branched goals, namely
switches and if-then-elses.  In those cases, it is still necessary to
unset the failure continuation, because if one or more of the branches is
model_nondet but the last branch is model_det or model_semidet, we
don't want to leave the failure continuation as it was before the
branched goal.

code_gen.pp:
	After generating a nondet goal, call code_info__unset_failure_cont.

disj_gen.m, call_gen.m:
	Remove calls to code_info__unset_failure_cont, since that is
	now done in code_gen.pp.

ite_gen.m:
	Before entering the condition, it pushes a new failure cont
	(pointing to the `else' case), and then after the condition, it
	pops this failure cont off again.  But if the condition is
	nondet, we need to make sure we set the top-level failure
	continuation to `unknown' before we generate the condition.
	Generating the condition will have done that, but it will
	have done it to the failure cont that we just popped!
	I've added a call to code_info__unset_failure_cont to do this
	to the newly exposed top of the failure cont stack.