Work around an installation problem on Irix.
Estimated hours taken: 0.5

Work around an installation problem on Irix.  When linking
`libmercury.so' in a non-gc grade, it linked against a `libmer.so'
that was compiled in a gc grade.  This made it search for `libgc.so',
but because there was no -L$(BOEHM_GC) option in the link command,
it couldn't find it.

library/Mmake:
	Always pass `-L$(BOEHM_GC)', even if linking a non-gc grade.