The performance of the compiler has been a concern recently.
Estimated hours taken: 3

The performance of the compiler has been a concern recently.
A little profiling and exprementation has shown that the
representation of sets as unordered lists with duplicats is
largely to blame. This commit adds a new module to the library
set_ordlist.m and makes set.m forward to it. It resulted in
about a 5%-10% speedup in the compilation of most modules, about
%10 slowdown for eliza (sigh) and about a 300% speedup on a
hand optimized version of the lexer.

library/library.m:
	add set_ordlist to the library.

library/set.m:
	forward to set_ordlist instead of set_unordlist.

library/set_ordlist.m:
	the new representation of lists as sorted lists with
	duplicates removed.