During the profiling signal handler it was possible that some memory would
Estimated hours taken: 0.5

During the profiling signal handler it was possible that some memory would
have to be malloc'd.  If the signal occured during another malloc then
internal state of malloc would get corrupted and hey presto seg fault.
Now memory for the profiler gets malloc'd in big chunks so to reduce the
chance of a malloc inside a malloc.

For a run of the compile now only do 4 malloc's instead of 6000.

profiler/prof_mem.*:
	The new memory management scheme for profiling.

profiler/prof.c:
	Changes to use the new scheme.

profiler/Mmake:
	Add the prof_mem.* stuff.