Fixed a bug in the C interface.
Estimated hours taken: 2.5

Fixed a bug in the C interface.
The names of C variables were indexed by the corresponding Mercury variable.
The problem arose when the same Mercury variable was passed in to a
C code in two separate arguments, as in string__append(X, X, Y).
If the procedure was inlined, excess assignments removed the assignment
introduced in make_hlds to put the clause into super-homogeneous form,
so a declaration for one C variable did not appear, and a declaration for
the other was repeated.


compiler/hlds_goal.m
compiler/*.m
	Change the method of storing C variable names from a map(var, string)
	to a list(maybe(string)), with one element for each argument to
	the C code.