Fix a bug that caused the generated executable for Tom's `strad'
Estimated hours taken: 6

Fix a bug that caused the generated executable for Tom's `strad'
program compiled in asm_fast.gc on the Alpha to get a segmentation
violation after returning from main/2.  The bug was due to the
return address being stored on the top of the stack and then
clobbered by some of the C code jumped to using non-local gotos.

runtime/engine.mod:
	Change the non-local gotos version of call_engine() so
	that it exits using longjmp() rather than by returning.