Move the `type_loop.m' test from test/valid to tests/invalid,
Estimated hours taken: 0.5

Move the `type_loop.m' test from test/valid to tests/invalid,
since the test code is in fact illegal, not legal, and the
compiler now correctly reports an error for it.

(The code in question is

	:- type foo == map(int, foo).

It would perhaps be reasonable to allow such code involving infinite
types, but currently we don't allow infinite types.  If we were to
change the language to allow infinite types, then we should do so
consistently everywhere.)

tests/invalid/{Mmake,type_loop.m}:
	Remove type_loop.m.