Fix a problem that caused link errors on Linux (which has mprotect, but
Estimated hours taken: 1.0

Fix a problem that caused link errors on Linux (which has mprotect, but
doesn't have siginfo). The fix lead to a bit of redesign in memory.{c,h}:
we don't need handler functions for catching accesses to redzones if we
don't have siginfo.

runtime/memory.{c,h}:
	Add #ifdefs to remove references to the handler functions if
	HAVE_SIGINFO is not defined.

	Move the stuff about PROT_* from memory.h to memory.c to avoid
	a compiler warning.