Run tests in all grades after installing the grades, so that the tests
Estimated hours taken: 3

Run tests in all grades after installing the grades, so that the tests
are run with the new libraries installed. This means a feature and its
test can be added on the same day, without giving a false failure of the
tests that night, followed by a successful test the next night (this was
the old behaviour, because tests were linked against old libraries).

tools/test_mercury:
	Move code for tests after installing, fix any problems this
	causes.
	Also remove a reference to a hardcoded directory for the
	install.$HOST file and successful_test.$HOST file.