Add support for generation position-independent code on SPARCs.
Estimated hours taken: 2

Add support for generation position-independent code on SPARCs.
(Currently we use non-PIC .so files, which means that although
they're shared on disk, they're not shared at runtime.)
Note that this change is not yet enabled -- the code I've added
will only be used if you compile with `-fpic', but the default
is still to compile without `-fpic'.

runtime/goto.h:
	Add code to set the global offset table pointer register
	(register `l7') to the correct value when entering or
	re-entering a procedure, when using PIC on SPARCs.

	Also add some comments to the code for doing the
	same thing for Intel x86s.