The bytecode file must store constants in Mercury programs in a
Estimated hours taken: 2

The bytecode file must store constants in Mercury programs in a
manner that is platform-independent. Specifically:
	- Mercury `int' constants are stored as 64-bit, 2's-complement,
	  big-endian quantities
	- Mercury `float' constants are stored as 64-bit, IEEE-754
	  big-endian floating point quantities.

The code that reads these quantities is neater if we can create
typedefs such as `int64' and `float64' that represent the above
entities.

To create such typedefs, we need sizeof information on C types.

configure.in
runtime/conf.h.in