It seems that when I added the runtests files for all the subdirectories
Estimated hours taken: 3

It seems that when I added the runtests files for all the subdirectories
in tests, I did not do a "cvs add" for invalid/runtests (even though the
log message says I did). As a consequence, our test runs since then
have not exercised the test cases in this directory. This checkin
fixes that oversight.

runtests:
	Shell script to exercise the test cases in this directory.

qual_basic_test2.*:
	A test case that was removed from hard_coded a while ago because
	it was really a test of the error message.

*.err_exp:
	Updated the expected error message outputs for several test cases
	according to the changes of the last few months. Added two missing
	.err_exp files.

Mmake:
	Disabled the tests prog_io_erroneous and qualified_cons_id2
	since we do not pass them, and I don't know what the expected output
	ought to be (it should not be what we output now).