Slight rearrangement of the data structures of the code generator, to allow
Estimated hours taken: 5

Slight rearrangement of the data structures of the code generator, to allow
more flexibility in code generation. The rearrangement moves the stack_slots
information (mapping vars to their stack slots if any) and follow_vars
information (mapping vars to the location preferred for them by future code)
from code_info to code_exprn. This allows the predicates in code_exprn to
make use of this information.

As a result of these changes, the code generator now emits 110 fewer lines
of C code for the compiler (478 lines are replaced by 368). There is no
discernible impact on the memory requirements or running time of the compiler
itself.

code_exprn:
	Add the two fields to the exprn_info data structure.

	Several predicates in code_exprn now evaluate variables directly
	into their preferred location, instead of a random register.

code_info:
	Remove the two fields from the code_info data structure.