Added a dictionary abstract data type. Currently it's implemented as
Estimated hours taken: 2

Added a dictionary abstract data type. Currently it's implemented as
a dirt-simple ordered list, but it should later be reimplemented using a
hash table. Note that C has only very poor support for ADTs
and private data must be included in the header. We trust clients
of the ADT not to use the private data.

bytecode/dict.h:
	Interface of dictionary ADT.
bytecode/dict.c:
	Implementation of dictionary ADT.