Fix a performance bug which meant that it took about 3 cpu minutes for
Estimated hours taken: 2

Fix a performance bug which meant that it took about 3 cpu minutes for
hydra to quote an 85K string in order to io__write it out. The bug was
that string__first_char(in, out, out) has a cost of O(n) in the length
of the string, and term_io__write_escaped_string called it successively
to escape each character of the string. The fix was to add a predicate
string__foldl which is equivalent to
	string__to_char_list(Str, Chars), list__foldl(Closure, Chars, Acc0, Acc)
except that its implementation avoids the intermediate list.

library/string.m:
	add string__foldl (described above)
	add the local predicate string__unsafe_index which doesn't do a
	range check to make sure that the index lies within the string.
	It is called by string__foldl.

library/term_io.m:
	implement term_io__write_escaped_string in terms of string__foldl
	instead of a loop calling string__first_char. The old version had
	a cost O(n^2), the new version a cost of O(n).

NEWS:
	mention string__foldl