Change the place in which the "#include" for a module's ".h" file is placed
Estimated hours taken: 2

Change the place in which the "#include" for a module's ".h" file is placed
when a procedure has been exported. Previously, the #include appeared just
before the definitions of the C functions for the exports. This change puts
the include at the top of the generated C file.

This has the consequence that the exported function can be used from
pragma c_code declarations without explicitly including the .h file.

compiler/llds_out.m:
        Don't bother outputting the include line when are outputing the
        exported functions.
compiler/mercury_compile.m:
        When we generate the pragma exports, add a "#include" to the
        c_header_info. This always gets put at the top of the generated file.

tests/hard_coded/export_test.*:
        A test for this bug fix.