Change the place in which the "#include" for a module's ".h" file is placed
Estimated hours taken: 2

Change the place in which the "#include" for a module's ".h" file is placed
when a procedure has been exported. Previously, the #include appeared just
before the definitions of the C functions for the exports. This change puts
the include at the top of the generated C file.

This has the consequence that the exported function can be used from
pragma c_code declarations without explicitly including the .h file.

tests/hard_coded/export_test.*:
        A test for this bug fix.