Fix a bug in lookup switch generation which meant that some semidet switches
Estimated hours taken: 6 (to track it down).

Fix a bug in lookup switch generation which meant that some semidet switches
were always succeeding. The problem was that, when the switch was dense enough
to get rid of the range check, getting rid of the range check meant that the
bit vector test could now fail.

It may be better to turn these on in the other order.

compiler/lookup_switch.m:
        When we get rid of the range check, turn the bit vector test on.