Print out a warning for clauses (etc.) in module interfaces.
Estimated hours taken: 1

Print out a warning for clauses (etc.) in module interfaces.
This was already done when creating the interface files,
but invalid code could slip past without a warning if
it occured in the topmost module (whose interface is not used,
and hence not created).

compiler/make_hlds.m:
	Print out a warning for clauses or certain pragmas (e.g.
	`pragma c_code') in module interfaces.
	Also inprove the maintainability of the code a little,
	by moving some code out into subroutines.

compiler/modules.m:
	Export pragma_allowed_in_interface/2, for use by make_hlds.m.