Disable the specialisation of class methods, since it did not work in the
Estimated hours taken: 1

Disable the specialisation of class methods, since it did not work in the
presence of implied instances. The problem is that where there is a constraint
on the instance decl, the methods for that instance need an extra typeclass
info, which did not get passed during specialisation.

compiler/polymorphism.m:
        Disable class method specialisation until a fix is found.