Fix the handling of the default linking options (shared or static).
Estimated hours taken: 0.5

Fix the handling of the default linking options (shared or static).

Under Linux, the default linkage is supposed to be static, but
only the linkage of the mercury libraries was being set, so
if you tried to link with another library, you needed to add
-static by hand.

Note that this will link in other libraries (such as the C library)
statically as well.

scripts/ml.in:
	Set link flags after handling options, so the flags are
	only set in one place.
	Add default handling for -static and -shared, previously
	there were only defaults for --mercury-libs static/shared.

README.Linux:
	Document that you should use --shared, not just --mercury-libs
	shared, if you want to link statically against other libraries.


Estimated hours taken: _____

<overview or general description of changes>

<directory>/<file>:
	<detailed description of changes>