Fix a bug that I introduced in modules.m.
Estimated hours taken: 0.5

Fix a bug that I introduced in modules.m.
The log message for my previous change was:

        - generate_dependencies was using the transitive implementation
	  dependencies rather than the transitive interface dependencies
	  to compute the `.int3' dependencies when writing `.d' files
	  (this bug was introduced during crs's changes to support
	  `.trans_opt' files)

Unfortunately using the transitive interface dependencies is not quite
right either.  You need to take the composition of the implementation
dependencies and the transitive interface dependencies.
For the moment I'll just back out this change, since the impact of the
previous bug was much less than the impact of the one I replaced it with.