Fix a problem that broke the use of NU-Prolog debugging.
Estimated hours taken: 0.25

Fix a problem that broke the use of NU-Prolog debugging.
The problem was that portray.nl was calling current_op/3,
but NU-Prolog only has currentOp/3.
The fix was to change things to use the NU-Prolog version
and add an implementation of the NU-Prolog version in
terms of the SICStus/ISO Prolog version to sp_lib.nl.
(A better policy would be always use the ISO names, and
to have an nc_lib.nl file providing implementations of
the ISO names in terms of the NU Prolog names rather than
vice versa; however, the current policy of always using
the NU-Prolog names is at least consistent, and changing
the current policy consistently would be a relatively large
amount of work.)

library/portray.nl:
	Change calls to current_op/3 to instead call currentOp/3.

library/sp_lib.nl:
	Add an implementation.