Fix a problem where the dependency files were being remade unnecessarily.
Estimated hours taken: 0.5

Fix a problem where the dependency files were being remade unnecessarily.
The problem was caused by the following change:

> Instead of deleting the `_init.c' file every time we recreate
> the `.dep' file, just tell make that the `_init.c' file depends on
> the `.dep' file, so that make will remake it if the `.dep' file
> changes.  (I don't know why I didn't do it that way in the
> first place.)

The reason we didn't want the `_init.c' file to define on the `.dep'
file is that the `.dep' file is listed as depending on the top-level
`.m' file.  The fix is to change things so that the `.dep' file
is not listed as depending on the `.m' file.

scripts/Mmake.rules:
	Fix some code duplication (and subsequent code rot, caused by
	changes in one place but not the other) in the rules for `.dep'
	and `.depend'.
	Delete the `.m' file from the dependencies for the `.dep' file.