Add an option --use-opt-files, which causes the compiler
Estimated hours taken: 2.5

Add an option --use-opt-files, which causes the compiler
to use any `.opt' files which are available when the
dependencies are made, but not make any more `.opt' files.
This is useful for using just the `.opt' files for the library
without building them for the compiler.

compiler/modules.m:
	When searching for available `.opt' files, don't look for
	`.m' files which could be used to build the `.opt' files
	if --use-opt-files is set.

compiler/handle_options.m:
compiler/options.m:
doc/user_guide.texi:
	Handle and document --use-opt-files.