Some changes to the compiler require that a user trying to build and install
Estimated hours taken: 1

Some changes to the compiler require that a user trying to build and install
a new version of the compiler from Mercury sources do the `mmake install'
step from the stage 2 directory.  This was not mentioned in the INSTALL_CVS;
this change fixes that omission.

INSTALL_CVS:
	Now runs a bootcheck by default and performs the `mmake install'
	from the stage 2 directory.  Updated the instructions to tell the
	user how to install without the bootcheck, and the implications of
	doing this.