There are 2 parts in this change (1) do not test whether or not an attribute is
Estimated hours taken: 30

There are 2 parts in this change (1) do not test whether or not an attribute is
wanted to retrieve it and (2) make the retrieval of arguments and other slots
done separately. The rationale for (1) is to reduce the average size of the
messages sent to the socket; for small fields, it takes less socket traffic to
always send the value, since it avoids the need to send a `bool' in the request
and a `maybe(T)' in the response. The rationale for (2) is that since
arguments slot can be very big, we still want to be able to retrieve any slot
without retrieving arguments; so we retrieve that slot separately.



library/debugger_interface.m:
	Split 'current' requests into 'current_var' and 'current_slots';
	'current_vars' requests for live variables and 'current_slots' for the
	other slots.

	output_current_slots/13 is a new predicate that sends to the debugger
	all the slots of the current event, except the argument slot.

	output_current_var/5 is a new predicate that sends to the debugger the
	list of currently live variables and the list of their internal names.

runtime/mercury_trace_external.c:
	Delete MR_output_current(); define MR_output_current_slots() and
	MR_output_current_vars().

	Define a function MR_trace_make_var_name_list() that returns the list
	of the internal name of the currently live variables.

runtime/mercury_init.h:
	Declare prototypes of ML_DI_output_current_vars and
	ML_DI_output_current_slots.

runtime/mercury_wrapper.h:
	Declare prototypes of MR_DI_output_current_vars and
	MR_DI_output_current_slots.

util/mkinit.c:
	Link MR_DI_output_current_vars to ML_DI_output_current_vars
	and MR_DI_output_current_slots to ML_DI_output_current_slots.