Phase out term__compare/4, since it is error-prone (the comparison
Estimated hours taken: 0.5

Phase out term__compare/4, since it is error-prone (the comparison
includes the term__contexts, which is probably not what you want)
and in general is probably not quite the right interface.
In addition, it's probably not of sufficient usefulness to
warrant inclusion in the standard library -- programmers can always
implement it themselves.

library/term.m:
	Add `pragma obsolete' declaration and some comments for
	term__compare/4, and move the declaration into a separate
	interface section so that it won't appear in the library
	reference manual.

NEWS:
	Mention this change.