This change alters the way `_' variables are printed.
Estimated hours taken: 0.1

This change alters the way `_' variables are printed. Currently they're
printed as _N nubered from zero by their first occurence when printed.
This means that if a term containing such variables is split and printed
using two separate calls to write_term, then the corresponding variables
may be printed differently. This change addresses this problem by using
an integer representation of the variable. This fixes the above problem,
but it will make terms printed as error messages a little harder to read,
for example what was printed:
	t(_0, list(T), _1, map(U, pair(V, W)))
may look like
	t(_23, list(T), _17, map(U, pair(V, W)))

The code for doing the first of these is still in place so that later
we can (if we decide to) add a flag to allow users to choose between
the two.

library/term_io.m:
	change term_io__write_variable_2 to use the variable's number
	for making a name for unnamed variables rather than using a
	counter that is local to the call to write_variable or write_term.